Se modifica la clase arma para que los atributos de id, pos_x, pos_y y alcance que recibe por parametro sean un uint16 para que se pueda mostrar en pantalla

+++ common/common_arma.cpp
@@ -1,6 +1,6 @@
 #include "common_arma.h"
 
-Arma::Arma(uint8_t id, uint8_t pos_x, uint8_t pos_y, uint8_t municion_disponible, uint8_t alcance):
+Arma::Arma(uint16_t id, uint16_t pos_x, uint16_t pos_y, uint8_t municion_disponible, uint16_t alcance):
         Entidad(id, pos_x, pos_y),
         municion_disponible(municion_disponible),
         alcance(alcance) {}
@@ -11,5 +11,5 @@ void Arma::disparar() { municion_disponible--; }
 
 uint8_t Arma::get_municion_disponible() { return municion_disponible; }
 
-uint8_t Arma::get_alcance() { return alcance; }
+uint16_t Arma::get_alcance() { return alcance; }
