+++ common/common_arma.h
@@ -8,16 +8,16 @@
 class Arma: public Entidad {
 protected:
     uint8_t municion_disponible;
-    uint8_t alcance;
+    uint16_t alcance;
 
 public:
-    Arma(uint8_t id, uint8_t pos_x, uint8_t pos_y, uint8_t municion_disponible,
-         uint8_t alcance);
+    Arma(uint16_t id, uint16_t pos_x, uint16_t pos_y, uint8_t municion_disponible,
+         uint16_t alcance);
     void disparar();
     void recargar();
     std::string get_nombre();
     uint8_t get_municion_disponible();
-    uint8_t get_alcance();
+    uint16_t get_alcance();
 };
 
 #endif
